+++ app/serializers/framework_assessment_serializer.rb
@@ -9,7 +9,7 @@ class FrameworkAssessmentSerializer
   has_many_if_included :responses, serializer: FrameworkResponseSerializer, &:framework_responses
   has_many_if_included :flags, serializer: FrameworkFlagSerializer, &:framework_flags
 
-  attributes :confirmed_at, :created_at, :nomis_sync_status
+  attributes :completed_at, :confirmed_at, :created_at, :nomis_sync_status
 
   attribute :version do |object|
     object.framework.version
@@ -24,7 +24,7 @@ class FrameworkAssessmentSerializer
   attribute :editable, &:editable?
 
   meta do |object|
-    { section_progress: object.calculate_section_progress }
+    { section_progress: object.section_progress }
   end
 
   SUPPORTED_RELATIONSHIPS = %w[