+++ app/serializers/person_escort_records_serializer.rb
@@ -8,9 +8,13 @@ class PersonEscortRecordsSerializer
 
   has_many_if_included :flags, serializer: FrameworkFlagsSerializer, &:framework_flags
 
-  attributes :confirmed_at, :created_at, :nomis_sync_status
+  attributes :created_at, :completed_at, :amended_at, :confirmed_at, :nomis_sync_status
 
   attribute :status do |object|
     object.status == 'unstarted' ? 'not_started' : object.status
+  end
+
+  meta do |object|
+    { section_progress: object.section_progress }
   end
 end