+++ app/services/locations/finder.rb
@@ -40,7 +40,9 @@ module Locations
     end
 
     def split_params(name)
-      filter_params[name]&.split(',')
+      return if filter_params[name].blank?
+
+      filter_params[name].split(',')
     end
 
     def apply_supplier_filters(scope)