+++ db/migrate/20210122172150_add_amended_at_to_person_escort_records.rb
@@ -1,0 +1,5 @@
+class AddAmendedAtToPersonEscortRecords < ActiveRecord::Migration[6.0]
+  def change
+    add_column :person_escort_records, :amended_at, :datetime
+  end
+end